backend/requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# HTTP client
httpx==0.27.0
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# JWT and Keycloak integration
python-jose[cryptography]==3.3.0

# Optional: SQLite database (usually built-in for Python)
